+++ internal/boot/boot_test.go
@@ -1346,6 +1346,7 @@ func defaultFullBootToolNames() []string {
 		"todo_write",
 		"wait",
 		"web_fetch",
+		"web_search",
 		"write_file",
 	}
 }
@@ -2158,7 +2159,7 @@ model = "x"
 func TestAddBuiltinsWithWorkspaceRootKeepsSessionTools(t *testing.T) {
 	reg := tool.NewRegistry()
 	var stderr bytes.Buffer
-	addBuiltins(reg, nil, []string{robustTempDir(t)}, sandbox.Spec{}, 120*time.Second, builtin.SearchSpec{}, &stderr, robustTempDir(t), netclient.ProxySpec{}, nil, nil, builtin.SessionDataGuard{})
+	addBuiltins(reg, nil, []string{robustTempDir(t)}, sandbox.Spec{}, 120*time.Second, builtin.SearchSpec{}, &stderr, robustTempDir(t), netclient.ProxySpec{}, nil, nil, builtin.SessionDataGuard{}, "", "", "")
 	for _, name := range []string{
 		"todo_write",
 		"complete_step",